pyproject.toml
[tool.poetry]
name = "selfspy"
version = "1.0.0"
description = "A tool for monitoring and analyzing your computer activity"
authors = ["nuin <nuin@genedrift.org>"]
readme = "README.md"
packages = [{include = "src"}]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0.25"
pynput = "^1.7.6"
cryptography = "^41.0.7"
keyring = "^24.3.0"
python-dateutil = "^2.8.2"
click = "^8.1.7"
appdirs = "^1.4.4"
rich = "^13.7.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.1.11"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
selfspy = "src.cli:main"
selfstats = "src.stats:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q --cov=src"